add delete functionality
