+++ src/index.html
@@ -11,7 +11,7 @@
         <h1>Today's To Do<span id="sync"><i class="bi bi-arrow-repeat"></i></span></h1>
         <form action="#">
             <textarea name="to-dos" id="add" placeholder="Add to your list..." rows="1" required></textarea>
-            <button type="submit"><i class="bi bi-arrow-return-left"></i></button>
+            <button id="return" type="submit"><i class="bi bi-arrow-return-left"></i></button>
         </form>
         <ul id="task-list"></ul>
         <button id= "clear" type="reset">Clear all completed</button>
